@@ -1,16 +1,15 @@
 package org.moire.opensudoku.utils;
 
-import java.util.List;
-
-import org.moire.opensudoku.R;
-
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
 import android.content.pm.PackageManager.NameNotFoundException;
+import android.content.pm.ResolveInfo;
 import android.preference.PreferenceManager;
+import org.moire.opensudoku.R;
+
+import java.util.List;
 
 public class AndroidUtils {
 	/**
@@ -36,24 +35,34 @@ public class AndroidUtils {
 	public static void setThemeFromPreferences(Context context) {
 		SharedPreferences gameSettings = PreferenceManager.getDefaultSharedPreferences(context);
 		String theme = gameSettings.getString("theme", "default");
-		if (theme.equals("default")) {
-			context.setTheme(R.style.Theme_Default);
-		} else if (theme.equals("paper")) {
-			context.setTheme(R.style.Theme_Paper);
-		} else if (theme.equals("graphpaper")) {
-			context.setTheme(R.style.Theme_GraphPaper);
-		} else if (theme.equals("light")) {
-			context.setTheme(R.style.Theme_Light);
-        } else if (theme.equals("paperlight")) {
-            context.setTheme(R.style.Theme_PaperLight);
-        } else if (theme.equals("graphpaperlight")) {
-            context.setTheme(R.style.Theme_GraphPaperLight);
-        } else if (theme.equals("highcontrast")) {
-            context.setTheme(R.style.Theme_HighContrast);
-        } else if (theme.equals("invertedhighcontrast")) {
-            context.setTheme(R.style.Theme_InvertedHighContrast);
-		} else {
-			context.setTheme(R.style.Theme_Default);
+		switch (theme) {
+			case "default":
+				context.setTheme(R.style.Theme_Default);
+				break;
+			case "paper":
+				context.setTheme(R.style.Theme_Paper);
+				break;
+			case "graphpaper":
+				context.setTheme(R.style.Theme_GraphPaper);
+				break;
+			case "dark":
+				context.setTheme(R.style.Theme_Dark);
+				break;
+			case "paperdark":
+				context.setTheme(R.style.Theme_PaperDark);
+				break;
+			case "graphpaperdark":
+				context.setTheme(R.style.Theme_GraphPaperDark);
+				break;
+			case "highcontrast":
+				context.setTheme(R.style.Theme_HighContrast);
+				break;
+			case "invertedhighcontrast":
+				context.setTheme(R.style.Theme_InvertedHighContrast);
+				break;
+			default:
+				context.setTheme(R.style.Theme_Default);
+				break;
 		}
 	}
 
